requirements.txt
pyaudio
SpeechRecognition
openai
google-api-python-client
google-cloud-speech==2.19.0
google-cloud-texttospeech==2.14.1
grpcio
python-voicetext
ffmpeg-python
python-dotenv
git+https://github.com/dsdanielpark/Bard-API
regex
websockets

types-pyaudio
types-regex
types-requests
types-beautifulsoup4
